Move Hosted Graphite python scripts to a testable directory
